Refactored RunTests method

+++ Assets/Scripts/PositionCounter.cs
@@ -5,11 +5,11 @@ using System.Diagnostics;
 
 public static class PositionCounter {
     public static int maxDepth;
-    public static int SearchPositions(GameState gameState, int depth, StreamWriter writer = null) {
+    public static long SearchPositions(GameState gameState, int depth, StreamWriter writer = null) {
         if (depth == maxDepth) {
             return 1;
         }
-        int sum = 0;
+        long sum = 0;
         Stopwatch stopwatch = new();
         stopwatch.Start();
         List<IndexMove> legalMoves = GetLegalMoves(gameState);
@@ -17,7 +17,7 @@ public static class PositionCounter {
         GameStateManager.Instance.numberOfTicks1 += stopwatch.ElapsedTicks;
         foreach (IndexMove move in legalMoves) {
             gameState.MakeMoveNoHashtable(move);
-            int numberPositions = SearchPositions(gameState, depth + 1, writer);
+            long numberPositions = SearchPositions(gameState, depth + 1, writer);
             if (depth == 0 && maxDepth == 1) {
                 writer?.WriteLine(new Move(move) + ": " + numberPositions);
             }
